backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==41.0.7
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
python-dotenv==1.0.0
werkzeug==3.0.1
